+++ test/block-kit/layout-blocks.tsx
@@ -27,7 +27,6 @@ import JSXSlack, {
   Section,
   Select,
   UsersSelect,
-  MrkDwn,
 } from '../../src/index'
 
 beforeEach(() => JSXSlack.exactMode(false))
@@ -240,65 +239,6 @@ describe('Layout blocks', () => {
       ).toStrictEqual([section]))
   })
 
-  describe('<Section> with <MrkDwn>', () => {
-    it('outputs a <Section> block wit a <MrkDwn> component and preserves the verbatim prop', () => {
-      expect(
-        JSXSlack(
-          <Blocks>
-            <Section>
-              <MrkDwn verbatim={false}>Hello!</MrkDwn>
-            </Section>
-          </Blocks>
-        )
-      ).toEqual([
-        {
-          type: 'section',
-          fields: [{ type: 'mrkdwn', text: 'Hello!', verbatim: false }],
-        },
-      ])
-    })
-
-    it('throws an error when <MrkDwn> contains an <img> element', () => {
-      expect(() =>
-        JSXSlack(
-          <Blocks>
-            <Section>
-              <MrkDwn verbatim={false}>
-                <img src="https://example.com/test.jpg" alt="Test image" />
-              </MrkDwn>
-            </Section>
-          </Blocks>
-        )
-      ).toThrow()
-    })
-
-    it('throws an error when <MrkDwn> contains an <Image> component', () => {
-      expect(() =>
-        JSXSlack(
-          <Blocks>
-            <Section>
-              <MrkDwn verbatim={false}>
-                <Image src="https://example.com/test.jpg" alt="Test image" />
-              </MrkDwn>
-            </Section>
-          </Blocks>
-        )
-      ).toThrow()
-    })
-
-    it('throws error when <Section> contains a <MrkDwn> child component and any other child element', () =>
-      expect(() =>
-        JSXSlack(
-          <Blocks>
-            <Section>
-              <MrkDwn verbatim={false}>Hello @user!</MrkDwn>
-              <span>Goodbye Error!</span>
-            </Section>
-          </Blocks>
-        )
-      ).toThrow())
-  })
-
   describe('<Divider>', () => {
     const divider: DividerBlock = {
       type: 'divider',
@@ -454,70 +394,6 @@ describe('Layout blocks', () => {
           ],
         },
       ]))
-
-    it('outputs a <Context> block with an image and multiple mrkdwn elements', () => {
-      expect(
-        JSXSlack(
-          <Blocks>
-            <Context>
-              <Image src="https://example.com/test.jpg" alt="Test Image" />
-              Supporting <code>context</code> block would be hard!
-              <MrkDwn verbatim={false}>
-                <i>@here</i>
-              </MrkDwn>
-              :dizzy_face:
-            </Context>
-          </Blocks>
-        )
-      ).toEqual([
-        {
-          type: 'context',
-          elements: [
-            {
-              type: 'image',
-              image_url: 'https://example.com/test.jpg',
-              alt_text: 'Test Image',
-            },
-            {
-              type: 'mrkdwn',
-              text: 'Supporting `context` block would be hard!',
-              verbatim: true,
-            },
-            {
-              type: 'mrkdwn',
-              text: '_@here_',
-              verbatim: false,
-            },
-            {
-              type: 'mrkdwn',
-              text: ':dizzy_face:',
-              verbatim: true,
-            },
-          ],
-        },
-      ])
-    })
-
-    it('converts <MrkDwn> elements into mrkdwn elements and preserves the verbatim prop', () =>
-      expect(
-        JSXSlack(
-          <Blocks>
-            <Context>
-              <MrkDwn verbatim={false}>Hello</MrkDwn>
-              World
-            </Context>
-          </Blocks>
-        )
-      ).toEqual([
-        {
-          type: 'context',
-          elements: [
-            { type: 'mrkdwn', text: 'Hello', verbatim: false },
-            { type: 'mrkdwn', text: 'World', verbatim: true },
-          ],
-        },
-      ]))
-
     it('throws error when the number of elements is 11', () =>
       expect(() =>
         JSXSlack(